Node editor improvements and some more stuff
NodeEditor Version 2:
Includes the clean version of what we had so far as well as some new checks that were missing.
The input and outputs are separated widgets now and all the logic is implemented through the blueprint interfaces.
Naming of the functions and variables has improved. The base node widget is setup as general node from which we can inherit more specific nodes.
Inputs and outputs can have a category now and a simple check and style is implemented for them already.